backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
fpdf==1.7.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
requests>=2.31.0
python-dotenv>=1.0.0